feat: refactor three component
