+++ src/assets/icons/outlined/Location.tsx
@@ -2,7 +2,7 @@ import { ISvgIcon } from "src/types/icons";
 
 const Location = ({color}: ISvgIcon) => (
     <svg width="22" height="22" viewBox="0 0 22 22" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path fill-rule="evenodd" clip-rule="evenodd" d="M10.6944 1.22223C6.48031 1.22223 3.05554 4.27778 3.05554 7.94445C3.05554 13.4689 10.6944 20.7778 10.6944 20.7778C10.6944 20.7778 18.3333 13.4689 18.3333 7.94445C18.3333 4.27778 14.9086 1.22223 10.6944 1.22223ZM10.6944 19.0056C8.32638 16.5245 4.32869 11.5378 4.32869 7.94445C4.32869 4.91334 7.19327 2.44445 10.6944 2.44445C12.4004 2.44445 14.0173 3.03112 15.2268 4.10667C16.3981 5.15778 17.0602 6.51445 17.0602 7.94445C17.0602 11.5378 13.0625 16.5245 10.6944 19.0056ZM13.2407 7.94445C13.2407 9.30112 12.1076 10.3889 10.6944 10.3889C9.28124 10.3889 8.14813 9.30112 8.14813 7.94445C8.14813 6.58779 9.28124 5.50001 10.6944 5.50001C12.1076 5.50001 13.2407 6.58779 13.2407 7.94445Z" fill={color}/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M10.6944 1.22223C6.48031 1.22223 3.05554 4.27778 3.05554 7.94445C3.05554 13.4689 10.6944 20.7778 10.6944 20.7778C10.6944 20.7778 18.3333 13.4689 18.3333 7.94445C18.3333 4.27778 14.9086 1.22223 10.6944 1.22223ZM10.6944 19.0056C8.32638 16.5245 4.32869 11.5378 4.32869 7.94445C4.32869 4.91334 7.19327 2.44445 10.6944 2.44445C12.4004 2.44445 14.0173 3.03112 15.2268 4.10667C16.3981 5.15778 17.0602 6.51445 17.0602 7.94445C17.0602 11.5378 13.0625 16.5245 10.6944 19.0056ZM13.2407 7.94445C13.2407 9.30112 12.1076 10.3889 10.6944 10.3889C9.28124 10.3889 8.14813 9.30112 8.14813 7.94445C8.14813 6.58779 9.28124 5.50001 10.6944 5.50001C12.1076 5.50001 13.2407 6.58779 13.2407 7.94445Z" fill={color}/>
     </svg>
 );
