+++ src/components/Breadcrumb/Breadcrumb.tsx
@@ -1,12 +1,12 @@
-import { FC } from "react";
+import { FC, Fragment } from "react";
 import * as Style from "./Breadcrumb.styles";
 import { IBreadcrumb } from "./Breadcrumb.types";
 
 const Breadcrumb: FC<IBreadcrumb> = ({breadcrumbs}) => (
     <Style.BreadcrumbList>
         {breadcrumbs.map((breadcrumb, index) => (
-            <>
-                <Style.BreadcrumbItem active={breadcrumb.active} key={breadcrumb.route}>
+            <Fragment key={`${breadcrumb.route}-${index}`}>
+                <Style.BreadcrumbItem $active={breadcrumb.active}>
                     <span>
                         {
                             (index > 0) && index < breadcrumbs.length && '/ '
@@ -14,7 +14,7 @@ const Breadcrumb: FC<IBreadcrumb> = ({breadcrumbs}) => (
                         {breadcrumb.label}
                     </span>
                 </Style.BreadcrumbItem>
-            </>
+            </Fragment>
         ))}
     </Style.BreadcrumbList>
 )
