+++ src/components/NavItem/NavItem.styles.ts
@@ -1,6 +1,6 @@
 import {styled} from "styled-components";
 
-const NavigationItem = styled.li<{active: boolean}>`
+const NavigationItem = styled.li<{$active: boolean}>`
     display: flex;
     align-items: center;
     justify-content: center;
@@ -8,7 +8,7 @@ const NavigationItem = styled.li<{active: boolean}>`
     padding: 4px 8px;
     font-size: 12px;
     font-weight: 600;
-    background-color: ${({ theme, active }) => active ? theme.colors["blue-500"] : theme.colors["blue-900"]};
+    background-color: ${({ theme, $active = false }) => $active ? theme.colors["blue-500"] : theme.colors["blue-900"]};
     color: #FFF;
     cursor: pointer;
 `;
