+++ src/components/ThreeList/ThreeList.style.ts
@@ -1,8 +1,8 @@
 import {styled} from "styled-components";
 
-const ThreeListContainer = styled.ul<{isChildren?: boolean}>`
+const ThreeListContainer = styled.ul<{$isChildren?: boolean}>`
     list-style: none;
-    padding: ${({isChildren = false}) => isChildren ? '4px 4px 4px 20px' : '0'};
+    padding: ${({$isChildren = false}) => $isChildren ? '4px 4px 4px 20px' : '0'};
     margin: 0;
 `;
